feat: TACLeBench paper-apps harness, PLL 110 MHz, FreeRTOS composability
Squashed batch covering four related work areas, all on STM32L552:

1. TACLeBench harness + paper-apps porting
   - tools/test_taclebench.sh: end-to-end OpenOCD+GDB harness with
     per-bench sequential passes, PHASE5_BENCHES=paper magic, CSV results,
     bash 3.x macOS-compatible, UART acquire safety, all-at-once mode.
   - host/taclebench/: blob_src wrappers + Makefile rules for kernel
     benches (recursion, fac, prime, binarysearch, bsort, countnegative)
     and sequential paper apps (ndes, statemate, dijkstra-disabled).
     libc_shim for memcpy/memset/EABI. Linker globs catch plain .rodata.
   - tools/protect_enclave.py: BFS-order block fold, gc-sections compatible,
     UMBRA_ESS_MISS_RECOVERY layout.
   - Preempt-loop fix (MPU AP=RW, ESS sizing, FAIL-path rollback).

2. Kernel improvements
   - UART-printing panic handler.
   - MAX_EFBS 32 → 64, BFS bitmap u32 → u64.
   - PSP stack 2K → 8K, MAX_ENCLAVES_CTX 4 → 2 (ndes fit).
   - CACHE_LIMIT_PER_ENCLAVE 24 → 64.
   - Force-load non-BFS blocks (replace evict-all).
   - efb_count = num_blocks before force-load.
   - MPCBB flipped to NS before DMA in handle_ess_miss.
   - DMA garbage-collect done request slots.

3. PLL @ 110 MHz + T-table AES (Stage A/C)
   - FLASH ACR (5WS + cache), PWR set_vos_range_boost, RCC HSI16+PLL.
   - platform_impl init_clocks canonical sequence, SysTick reload, USART1
     routed to HSI16 on L562.
   - OCTOSPI prescaler /3 → /8 for 110 MHz SYSCLK reliability (L562).
   - _SECURE_BOOT_TEXT_MEMORY_ 68K → 72K for L562 build.
   - T-table AES encrypt_block in AesEmulated on L552 and N657
     (~8× speedup; statemate 88s → 5s pre-FreeRTOS).

4. FreeRTOS paper-apps composability host
   - host/freertos_arm/src/main.c: vHeartbeatTask (prio 2, 1KB stack,
     vTaskDelay 100ms) + drift_stats_t with 6-bucket histogram + DWT.CYCCNT
     vApplicationTickHook + end-of-run [DRIFT] dump. taskENTER_CRITICAL
     wraps NS-side UART prints to prevent NS_SysTick from preempting
     Secure thread mid-print; not wrapped around umbra_enclave_enter
     (heartbeats must fire during enclave execution) nor around
     umbra_enclave_create (drift bound trade-off).
   - host/freertos_arm/src/FreeRTOSConfig.h: configCPU_CLOCK_HZ 4M → 110M,
     configTOTAL_HEAP_SIZE 32K → 16K, configUSE_TICK_HOOK = 1.
   - tools/test_taclebench.sh extended HOST_APP-aware: derives
     HOST_LOG_PREFIX from HOST_APP env (settings.sh), parametrizes
     "NS world alive" check, adds freertos-only assertions (heartbeat
     count ≥ 5, drift max < 6_600_000 cycles ≈ 60ms, healthy bucket
     fraction b0..b3 ≥ 70%). Extended CSV to 9 columns. PHASE5_ALL_AT_ONCE
     + HOST_APP=freertos die-s early (drift assertions are bench-specific).
     Bash double-quoted echo backtick → command-sub bug fix.

Hardware validation (L552):
- HOST_APP=bare_metal PHASE5_BENCHES=paper → 5/5 PASS (binarysearch,
  bsort, countnegative, ndes, statemate). Wall times 4-5s/bench.
- HOST_APP=freertos PHASE5_BENCHES=paper → 5/5 PASS, same R0 values,
  wall times 5s/bench (≤1.5× bare_metal baseline). Per bench:
  heartbeat_count=26, drift_max ≈ 5.16M cycles (47ms), healthy_pct 98-99%.

+++ src/hardware/platform/stm32l552/boot/src/prefetch.rs
@@ -18,7 +18,10 @@ pub unsafe fn prefetch_reachables(enclave_id: u32) {
     let mut queue = [0u8; QUEUE_SIZE];
     let mut head: usize = 0;
     let mut tail: usize = 0;
-    let mut visited: u32 = 0;
+    // Bitmap for visited blocks. MUST be wide enough for MAX_EFBS bits.
+    // u64 covers 64 — matches the kernel's BFS-state width in
+    // api_impl.rs umbra_enclave_create_imp.
+    let mut visited: u64 = 0;
 
     // Seed: enqueue reachables of all currently-loaded blocks
     {
@@ -28,14 +31,14 @@ pub unsafe fn prefetch_reachables(enclave_id: u32) {
         };
         for i in 0..enclave.efb_count {
             if enclave.efbs[i].is_loaded {
-                visited |= 1 << i;
+                visited |= 1u64 << i;
                 for r in 0..enclave.efbs[i].reachable_count as usize {
                     let target = enclave.efbs[i].reachable[r] as usize;
-                    if target < enclave.efb_count && (visited & (1 << target)) == 0 {
+                    if target < enclave.efb_count && (visited & (1u64 << target)) == 0 {
                         if !enclave.efbs[target].is_loaded && tail < QUEUE_SIZE {
                             queue[tail] = target as u8;
                             tail += 1;
-                            visited |= 1 << target;
+                            visited |= 1u64 << target;
                         }
                     }
                 }
@@ -88,13 +91,13 @@ pub unsafe fn prefetch_reachables(enclave_id: u32) {
             for r in 0..efb.reachable_count as usize {
                 let target = efb.reachable[r] as usize;
                 if target < enclave.efb_count
-                    && (visited & (1 << target)) == 0
+                    && (visited & (1u64 << target)) == 0
                     && !enclave.efbs[target].is_loaded
                     && tail < QUEUE_SIZE
                 {
                     queue[tail] = target as u8;
                     tail += 1;
-                    visited |= 1 << target;
+                    visited |= 1u64 << target;
                 }
             }
         }
